Separate SymmetricMonoidal to its own module
