+++ linear-base.cabal
@@ -20,6 +20,7 @@ library
     Control.Monad.Builder
     Control.Monad.Linear
     Control.Monad.Linear.Builder
+    Data.Bifunctor.Linear
     Data.Functor.Linear
     Data.Vector.Linear
     Data.Profunctor.Linear
